feat(parsing): add 'assert' notation
This notation works like an if but with a lower precedence. It will
check the given condition and wrap the whole remaining block into the
'true'-branch, thus making sure the condition holds for the current
block. A block can be understood intuitively as everything that may
possibly belong to the assert. A block only ends if something is closed
that was started before the assert e.g. a paranthesis or a match arm
end.

This precedence gives the assert the ability to de-nest code as the
'true'-block after the condition does not have to be indented, making
code more redable.