Vect.v: rename scope 'vect' to 'vect_scope'
According to the manual, scope names should end with a '_scope' prefix.
Thus I adapted this best practice for vect.

Additionally this commit add a scope binding to ease the use of vect's.
This will cause coq to automatically interpret these list notations as
vectors wherever a 'vect' type is expected. Thus eliminating the need
for most of the '%vect' annotations.